// frontend/src/types/appWidget.ts
export interface IAppWidget {
  id: string;
  createdAt: string;
  updatedAt: string;
  agentName: string;
  agentPhoto?: string 
  headerColor: string;
  headerTextColor: string;
  agentMessageColor: string;
  agentTextColor: string;
  visitorMessageColor: string;
  visitorTextColor: string;
  appId: string;
}